[chore] extract express app
